+++ backend/crypto/app.py
@@ -1,0 +1,9 @@
+from flask import Flask
+
+
+
+app = Flask(__name__)
+
+
+
+from crypto import routes
